CS-107 fix: format bigdecimal value of money instead of double

+++ types/src/main/java/ish/math/format/PlainMoneyDecimalFormatter.java
@@ -39,6 +39,6 @@ public class PlainMoneyDecimalFormatter implements MoneyFormatter {
 			symbols.setDecimalSeparator('.');
 			((DecimalFormat) xmlDecimal).setDecimalFormatSymbols(symbols);
 		}
-		return xmlDecimal.format(money.toDouble());
+		return xmlDecimal.format(money.toBigDecimal());
 	}
 }
